+++ spring-microservices/course-service/src/main/java/com/hoangphihiep/repository/CourseEnrollmentRepository.java
@@ -12,65 +12,65 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface CourseEnrollmentRepository extends JpaRepository<CourseEnrollment, Long> {
+public interface CourseEnrollmentRepository extends JpaRepository<CourseEnrollment, Integer> {
 
     // Class-based enrollment methods
-    Page<CourseEnrollment> findByCourseClassId(Long classId, Pageable pageable);
+    Page<CourseEnrollment> findByCourseClassId(Integer classId, Pageable pageable);
 
     Page<CourseEnrollment> findByStudentId(String studentId, Pageable pageable);
 
-    Optional<CourseEnrollment> findByCourseClassIdAndStudentId(Long classId, String studentId);
+    Optional<CourseEnrollment> findByCourseClassIdAndStudentId(Integer classId, String studentId);
 
-    boolean existsByCourseClassIdAndStudentId(Long classId, String studentId);
+    boolean existsByCourseClassIdAndStudentId(Integer classId, String studentId);
 
-    int countByCourseClassId(Long classId);
+    int countByCourseClassId(Integer classId);
 
     @Query("SELECT ce.studentId FROM CourseEnrollment ce WHERE ce.courseClass.id = :classId")
-    List<String> findStudentIdsByClassId(@Param("classId") Long classId);
+    List<String> findStudentIdsByClassId(@Param("classId") Integer classId);
 
     @Query("SELECT ce FROM CourseEnrollment ce WHERE ce.courseClass.id = :classId AND ce.status = :status")
-    List<CourseEnrollment> findByCourseClassIdAndStatus(@Param("classId") Long classId, @Param("status") String status);
+    List<CourseEnrollment> findByCourseClassIdAndStatus(@Param("classId") Integer classId, @Param("status") String status);
 
-    void deleteByCourseClassIdAndStudentId(Long classId, String studentId);
+    void deleteByCourseClassIdAndStudentId(Integer classId, String studentId);
 
     // Legacy methods for backward compatibility (can be removed if not used elsewhere)
     @Deprecated
-    Page<CourseEnrollment> findByCourseId(int courseId, Pageable pageable);
+    Page<CourseEnrollment> findByCourseId(Integer courseId, Pageable pageable);
 
     @Deprecated
-    Optional<CourseEnrollment> findByCourseIdAndStudentId(int courseId, String studentId);
+    Optional<CourseEnrollment> findByCourseIdAndStudentId(Integer courseId, String studentId);
 
     @Deprecated
-    boolean existsByCourseIdAndStudentId(int courseId, String studentId);
+    boolean existsByCourseIdAndStudentId(Integer courseId, String studentId);
 
     @Deprecated
-    int countByCourseId(int courseId);
+    int countByCourseId(Integer courseId);
 
     @Deprecated
     @Query("SELECT ce.studentId FROM CourseEnrollment ce WHERE ce.course.id = :courseId")
-    List<String> findStudentIdsByCourseId(@Param("courseId") int courseId);
+    List<String> findStudentIdsByCourseId(@Param("courseId") Integer courseId);
 
     @Deprecated
     @Query("SELECT ce FROM CourseEnrollment ce WHERE ce.course.id = :courseId AND ce.status = :status")
-    List<CourseEnrollment> findByCourseIdAndStatus(@Param("courseId") int courseId, @Param("status") String status);
+    List<CourseEnrollment> findByCourseIdAndStatus(@Param("courseId") Integer courseId, @Param("status") String status);
 
     @Deprecated
-    void deleteByCourseIdAndStudentId(int courseId, String studentId);
+    void deleteByCourseIdAndStudentId(Integer courseId, String studentId);
 
     // Convenience methods for class-based operations
-    default Page<CourseEnrollment> findByClassId(Long classId, Pageable pageable) {
+    default Page<CourseEnrollment> findByClassId(Integer classId, Pageable pageable) {
         return findByCourseClassId(classId, pageable);
     }
 
-    default boolean existsByClassIdAndStudentId(Long classId, String studentId) {
+    default boolean existsByClassIdAndStudentId(Integer classId, String studentId) {
         return existsByCourseClassIdAndStudentId(classId, studentId);
     }
 
-    default int countByClassId(Long classId) {
+    default int countByClassId(Integer classId) {
         return countByCourseClassId(classId);
     }
 
-    default void deleteByClassIdAndStudentId(Long classId, String studentId) {
+    default void deleteByClassIdAndStudentId(Integer classId, String studentId) {
         deleteByCourseClassIdAndStudentId(classId, studentId);
     }
 }
